requirements.txt
numpy==1.26.4
scipy==1.12.0
pandas==2.2.2
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
matplotlib==3.8.4
transformers==4.40.0
tqdm==4.66.2
scikit-learn==1.4.2
scikit-image==0.23.2
IPython==8.23.0
nltk==3.8.1
nbconvert==7.16.3
gdown==5.1.0
ipykernel==6.29.4
sacremoses==0.1.1
gensim==4.3.2